+++ CHANGELOG.md
@@ -11,7 +11,6 @@ _The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/)
 
 - Handle situation if stale elements detected when cookie cannot be loaded
 - Convert Documentation to Docusaurus
-- Check if user's account is private
 
 ### Fixed
 
@@ -19,6 +18,7 @@ _The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/)
 - Fixed the problem where `followers_list` could be used without being initialized.
 - Fixed incorrect xpath used by `interact_by_comments`
 - Fixed typo in `dissmiss_save_information` to `dismiss_save_information`
+- Reverted check if user's account is private
 
 ## [0.6.12] - 2020-10-26